backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
unicorn==2.1.3
uvicorn==0.34.2
